Create `FilesContextProvider` component to share instantiation logic
